pyproject.toml
[tool.poetry]
name = "django-backend"
version = "0.1.0"
description = ""
authors = ["jokwangjo89 <samgugjojo@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.4"
paho-mqtt = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
